优化接口实现

+++ trade-service/src/main/java/com/campus/trade/trade/dto/CreateTradeRequest.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.awt.geom.Arc2D;
 import java.math.BigDecimal;
 
 public class CreateTradeRequest {
@@ -13,7 +14,7 @@ public class CreateTradeRequest {
 
     @NotNull(message = "price不能为空")
     @DecimalMin(value = "0.0", inclusive = false, message = "price必须大于0")
-    private BigDecimal price;
+    private Double price;
 
     @NotBlank(message = "sellerId不能为空")
     private String sellerId;
@@ -26,11 +27,11 @@ public class CreateTradeRequest {
         this.itemId = itemId;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
